Improvements to the near_field package.
1) Issue 123 describes the inconsistency between ADDA and near_field. ADDA's issue was corrected by r1008, but there still exist some non-trivial functionality in ADDA which can lead to weird shape files. near_field's issue (when working with weird shape files) can be partly alleviated by simple suggestion given by Fabio http://groups.google.com/group/adda-discuss/msg/a1110e72866576fd . 

However, the limitation of it is the assumption that center of the particle coincides (within shift of 0.5 dipoles) with the center of the computational domain. Moreover, two variable (rc{x,y,z} and ic{x,y,z}) are used to describe almost the same quantity, while difference between them is used as an indicator of parity of number of dipoles along the corresponding axes.

The correction sets rc equal to exact center of the computational domain, using both target (integer coordinates) and DipPol (real dipole coordinates). Thus rc is further directly used in e.g. calcbeam to get real coordinates of any point. ic was replaced by pc, which is double the coordinates of the particle center, i.e. the center of the _minimal_ box (with possibly odd sizes) enclosing the particle. Since this center is either integer or half-integer, pc is always integer, which is convenient for further use.

The new version of near_field should work with any shape files, even those not-completely-correct produced by older ADDA versions. This was checked on data provided at http://code.google.com/p/a-dda/issues/detail?id=123#c8 (with small modifications).

2) One of the ways to produce weird shape files, for which the above correction is necessary, is to run ADDA with "-shape read ... grid ..." where arguments to grid are larger than coordinates sizes given by the shape file. Thus a test case was added to RUNTESTS/readodd, which should not be passed by previous version of near_field.

3) All 'run...' scripts were improved by adding checking of ADDA errors. So now the test will not continue if ADDA fails. Also added 'clean' scripts to remove all files produced during runs.