A separate branch - proof-of-principle implementation of the weighted discretization. It works only for spheres without anisotropy. Includes the following features (they may be useful in other applications as well) 1) Continuous refractive index and polarizability (increases memory requirements a little bit). 2) Simplest version of weighted discretization by using effective refractive index, determined by the occupied volume fraction of the dipole. 3) Routine to approximately determine the volume of intersection of the cube with the sphere. It starts with replacing a sphere with a tangent plane, which error is order of d/R (d - dipole size, R - radius of curvature). This simple approximation can be easily generalized to other shapes. Additionally, sphere-specific formulae are derived for accurate correction of order d/R, so the final inaccuracy is of order (d/R)^2. However, the latter corrections are not so straightforward to generalize. 4) Routine AbsOutProd in cmplx.h for absolute value of the outer product of two vectors.
Limited tests of this implementation were performed. The total volume of all dipoles (judged by the volume-equivalent radius) is very close to the true value. Moreover, the difference (error) goes to zero with order of d/R between 3 and 4 (which is as it should).

The accuracy of Cext is better than with standard discretization, but approximately the same as when using the simple volume correction of the whole dipole set.

To conclude, a lot of further development and research is required to understand the benefits of using weighting discretization. In particular, the weighting by Piller and Martin need to be implemented.