Version incremented to 1.1b5.
Win32 executables were updated, as well as corresponding files in sample/. Four DLLs were added to Win32 folder - they are required to run Fortran and C++ code, compiled with gcc.  Alternative option is to use static linking, but considering the number of executables this is expected to be less efficient. 'win32/README' added to explain that DLLs are also needed for executables in 'win32/misc'.