Implemented calculations of decay-rate enhancement for point-dipole incident beam. Total part is calculated with the help of a new function DecayCross() in crosssec.c/h (similar to Cext), non-radiative part - by normalizing Cabs. In case of surface, all parts of decay-rate enhancement (total, rad, nonrad) are normalized to that in free space, but additionally enhancement due to surface alone is shown.
Still, it is not clear whether the non-radiative part should be adjusted to account for absorption in the substrate (when it is absorbing).

Also, incident polarizations (Y and X) are not shown in the log for point-dipole incident beam. This is a bit problematic for standard LDR polarizability, since its factor S depends on both prop and incPol. But it probably makes little sense to use LDR polarizability for such incident field anyway (recommendation should be added to the manual).

One test was run for dipole near a sphere was run and matched exactly to first row of Table 1 in S. D'Agostino, F. Della Sala, and L.C. Andreani, “Dipole-excited surface plasmons in metallic nanoparticles: Engineering decay dynamics within the discrete-dipole approximation,” Phys. Rev. B 87, 205413 (2013). However, a few hacks were used to exactly match the simulation conditions (based on raw data provided by Stefania D'Agostino.